# catalogo/specs/001-catalogo-sessoes-reserva/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Servico-Catalogo API
  version: 1.0.0
  description: |
    API pública de navegação do catálogo de cinema e ponte síncrona para a reserva
    de poltronas no Servico-Estoque.

    Duas escolhas que valem para toda a API:
      1. Coleções respondem com envelope de paginação, não com array nu.
      2. Erros seguem RFC 9457 (application/problem+json), inclusive o 409.
servers:
  - url: /api/v1

# A ordem desta lista é a ordem de exibição na documentação: navegação primeiro,
# na sequência em que o usuário percorre o catálogo, e operação por último.
...
paths:
  /filmes:
    get:
      tags: [Filmes]
      summary: Lista filmes do catálogo
      description: |
        Sem o parâmetro `status`, retorna apenas filmes EM_CARTAZ e BREVE.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [EM_CARTAZ, BREVE, FORA_DE_CARTAZ]
      responses:
        '200':
          description: Página de filmes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginaFilmes' }
        '400': { $ref: '#/components/responses/EntradaInvalida' }
    post:
      tags: [Filmes]
      summary: Cadastra um filme
      description: |
        O identificador é gerado pelo serviço e devolvido no corpo e no header
        `Location`. Sem `status` no corpo, o filme nasce EM_CARTAZ.
      security:
        - clientCredentials: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FilmeEntrada' }
      responses:
        '201':
          description: Filme cadastrado
          headers:
            Location:
              description: Caminho do filme recém-criado
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Filme' }
        '400': { $ref: '#/components/responses/EntradaInvalida' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
...
  /cinemas:
    get:
      tags: [Cinemas]
      summary: Lista cinemas
      description: |
        Sem o parâmetro `ativo`, retorna apenas os cinemas ativos.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: ativo
          in: query
          required: false
          schema: { type: boolean }
      responses:
        '200':
          description: Página de cinemas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginaCinemas' }
        '400': { $ref: '#/components/responses/EntradaInvalida' }
    post:
      tags: [Cinemas]
      summary: Cadastra um cinema
      description: |
        O identificador é gerado pelo serviço e devolvido no corpo e no header
        `Location`. Sem `ativo` no corpo, o cinema nasce ativo.
      security:
        - clientCredentials: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CinemaEntrada' }
      responses:
        '201':
          description: Cinema cadastrado
          headers:
            Location:
              description: Caminho do cinema recém-criado
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cinema' }
        '400': { $ref: '#/components/responses/EntradaInvalida' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
...
  /sessoes:
    get:
      tags: [Sessões]
      summary: Consulta a grade de sessões
      description: |
        Retorna apenas sessões AGENDADA ou EM_ANDAMENTO.
        Filtros podem ser combinados.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: filme_id
          in: query
          schema: { type: string, format: uuid }
        - name: cinema_id
          in: query
          schema: { type: string, format: uuid }
        - name: data
          in: query
          description: Data no formato YYYY-MM-DD
          schema: { type: string, format: date }
      responses:
        '200':
          description: Página da grade de sessões
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginaSessoes' }
        '400': { $ref: '#/components/responses/EntradaInvalida' }
    post:
      tags: [Sessões]
      summary: Publica uma sessão na grade
      description: |
        O identificador é gerado pelo serviço e devolvido no corpo e no header
        `Location`. O filme e a sala precisam existir, e a sala precisa estar
        livre no intervalo que vai do início até o fim do filme: uma sala
        projeta um filme de cada vez.
      security:
        - clientCredentials: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessaoEntrada' }
      responses:
        '201':
          description: Sessão publicada
          headers:
            Location:
              description: Caminho da sessão recém-criada
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sessao' }
        '400': { $ref: '#/components/responses/EntradaInvalida' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '404': { $ref: '#/components/responses/NaoEncontrado' }
        '409': { $ref: '#/components/responses/Conflito' }
...
components:
  securitySchemes:
    clientCredentials:
      type: oauth2
      description: |
        Fluxo máquina-a-máquina: a service account do client troca
        client_id/client_secret por um token, sem usuário humano no meio.
        Use `cinema-m2m` / `segredo-de-desenvolvimento`.

        O `tokenUrl` aponta para `localhost` porque quem o chama é o navegador,
        não o serviço. Use este fluxo pelo Swagger apenas em desenvolvimento: o
        secret trafega pelo navegador e o client deixa de ser confidencial.
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8081/realms/cinema/protocol/openid-connect/token
          scopes: {}
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      required: false
      description: |
        Teto configurável por ambiente (`PAGINACAO_TAMANHO_MAXIMO`, padrão 100).
        Acima do teto vigente a requisição é recusada com 400 e
        `type: .../parametro-invalido`, cujo `detail` informa o máximo aceito.
      schema: { type: integer, minimum: 1, default: 20 }
  schemas:
    Paginacao:
      type: object
      required: [pagina, tamanho, total, tem_proxima]
      properties:
        pagina: { type: integer, example: 1 }
        tamanho: { type: integer, example: 20 }
        total:
          type: integer
          description: Total de registros que atendem aos filtros
          example: 137
        tem_proxima: { type: boolean, example: true }
    Filme:
      type: object
      required: [id, titulo, duracao_minutos, classificacao_etaria, genero, status]
      properties:
        id: { type: string, format: uuid }
        titulo: { type: string }
        sinopse: { type: string, description: Omitido quando ausente }
        duracao_minutos: { type: integer }
        classificacao_etaria: { type: string }
        genero: { type: string }
        imagem_url: { type: string, format: uri, description: Omitido quando ausente }
        status: { type: string, enum: [EM_CARTAZ, BREVE, FORA_DE_CARTAZ] }
    FilmeEntrada:
      type: object
      description: Corpo de escrita de um filme; o `id` nunca vem do cliente.
      required: [titulo, duracao_minutos, classificacao_etaria, genero]
      properties:
        titulo: { type: string, minLength: 1, maxLength: 255 }
        sinopse: { type: string }
        duracao_minutos: { type: integer, minimum: 1 }
        classificacao_etaria: { type: string, minLength: 1, maxLength: 50 }
        genero: { type: string, minLength: 1, maxLength: 100 }
        imagem_url: { type: string, format: uri, maxLength: 500 }
        status:
          type: string
          enum: [EM_CARTAZ, BREVE, FORA_DE_CARTAZ]
          default: EM_CARTAZ
      additionalProperties: false
    Cinema:
      type: object
      required: [id, nome, cidade, estado, endereco, ativo]
      properties:
        id: { type: string, format: uuid }
        nome: { type: string }
        cidade: { type: string }
        estado: { type: string, minLength: 2, maxLength: 2 }
        endereco: { type: string }
        ativo: { type: boolean }
    CinemaEntrada:
      type: object
      description: Corpo de escrita de um cinema; o `id` nunca vem do cliente.
      required: [nome, cidade, estado, endereco]
      properties:
        nome: { type: string, minLength: 1, maxLength: 255 }
        cidade: { type: string, minLength: 1, maxLength: 100 }
        estado:
          type: string
          minLength: 2
          maxLength: 2
          description: Sigla da unidade federativa, normalizada para maiúsculas
        endereco: { type: string, minLength: 1 }
        ativo: { type: boolean, default: true }
      additionalProperties: false
...
    Sessao:
      type: object
      description: |
        A sessão como é gravada. A grade devolve `SessaoNaGrade`, com o filme e o
        cinema já resolvidos.
      required: [id, filme_id, sala_id, data_hora_inicio, idioma, preco_base, status]
      properties:
        id: { type: string, format: uuid }
        filme_id: { type: string, format: uuid }
        sala_id: { type: string, format: uuid }
        data_hora_inicio: { type: string, format: date-time }
        idioma: { type: string, enum: [DUBLADO, LEGENDADO] }
        preco_base:
          type: string
          description: Decimal exato em texto, para não perder precisão em ponto flutuante
          example: '42.00'
        status: { type: string, enum: [AGENDADA, EM_ANDAMENTO, FINALIZADA, CANCELADA] }
    SessaoEntrada:
      type: object
      required: [filme_id, sala_id, data_hora_inicio, idioma, preco_base]
      properties:
        filme_id: { type: string, format: uuid }
        sala_id: { type: string, format: uuid }
        data_hora_inicio: { type: string, format: date-time }
        idioma: { type: string, enum: [DUBLADO, LEGENDADO] }
        preco_base:
          type: string
          description: Decimal exato em texto, com até duas casas
          example: '42.00'
        status:
          type: string
          enum: [AGENDADA, EM_ANDAMENTO, FINALIZADA, CANCELADA]
          default: AGENDADA
    SessaoNaGrade:
      type: object
      required: [id, filme_id, filme_titulo, cinema_id, cinema_nome, sala_numero, tipo_tela, data_hora_inicio, idioma, preco_base]
      properties:
        id: { type: string, format: uuid }
        filme_id: { type: string, format: uuid }
        filme_titulo: { type: string }
        cinema_id: { type: string, format: uuid }
        cinema_nome: { type: string }
        sala_numero: { type: integer }
        tipo_tela: { type: string, enum: ['2D', '3D', IMAX, VIP] }
        data_hora_inicio: { type: string, format: date-time }
        idioma: { type: string, enum: [DUBLADO, LEGENDADO] }
        preco_base:
          type: string
          description: Decimal exato em texto, para não perder precisão em ponto flutuante
          example: '42.00'
    PaginaFilmes:
      type: object
      required: [itens, pagina]
      properties:
        itens: { type: array, items: { $ref: '#/components/schemas/Filme' } }
        pagina: { $ref: '#/components/schemas/Paginacao' }
    PaginaCinemas:
      type: object
      required: [itens, pagina]
      properties:
        itens: { type: array, items: { $ref: '#/components/schemas/Cinema' } }
        pagina: { $ref: '#/components/schemas/Paginacao' }
...
    PaginaSessoes:
      type: object
      required: [itens, pagina]
      properties:
        itens: { type: array, items: { $ref: '#/components/schemas/SessaoNaGrade' } }
        pagina: { $ref: '#/components/schemas/Paginacao' }
...
    Problem:
      type: object
      description: RFC 9457
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              campo: { type: string }
              mensagem: { type: string }
  responses:
...
    EntradaInvalida:
      description: Parâmetro ou corpo inválido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NaoAutenticado:
      description: Credencial ausente, inválida, expirada ou sem identificação do usuário
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NaoEncontrado:
      description: Recurso inexistente
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflito:
      description: |
        Conflito com o estado atual: poltrona já tomada, número de sala repetido
        ou sala ocupada no horário.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...